Refactor folder structure for design pattern
