+++ packages/client/gateways/users_gateway.ts
@@ -1,7 +1,7 @@
 import { JwtToken } from "@cryptify/common/src/domain/jwt_token";
 import { User } from "@cryptify/common/src/domain/entities/user";
 import { ApiGateway } from "./api_gateway";
-import { Method } from "@cryptify/common/src/gateway/abstract_gateway";
+import { Method } from "@cryptify/common/src/utils/gateway/abstract_gateway";
 
 export class UsersGateway extends ApiGateway {
     constructor() {
