+++ packages/common/src/utils/currency_utils.ts
