+++ packages/common/src/utils/function_utils.ts
