+++ packages/common/src/utils/gateway/abstract_gateway.ts
