+++ packages/common/src/utils/string_utils.ts
