+++ packages/eth-edge/src/gateways/alchemy_node.gateway.ts
@@ -1,7 +1,7 @@
 import { Injectable } from "@nestjs/common";
 import { ConfigService } from "@nestjs/config";
-import { Gateway } from "@cryptify/common/src/gateway/gateway";
-import { Method } from "@cryptify/common/src/gateway/abstract_gateway";
+import { Gateway } from "@cryptify/common/src/utils/gateway/gateway";
+import { Method } from "@cryptify/common/src/utils/gateway/abstract_gateway";
 
 @Injectable()
 export class AlchemyNodeGateway extends Gateway {
